+++ prince-archiver/prince_archiver/worker.py
@@ -1,15 +1,12 @@
 import asyncio
 import logging
 import os
-import tarfile
-from concurrent.futures import Executor, ProcessPoolExecutor
-from contextlib import AsyncExitStack, asynccontextmanager
+from concurrent.futures import ProcessPoolExecutor
+from contextlib import AsyncExitStack
 from pathlib import Path
-from typing import AsyncGenerator
 
 import aiofiles
 import aiofiles.os
-import cv2
 import s3fs
 from aiofiles.tempfile import TemporaryDirectory
 from arq.connections import RedisSettings
@@ -17,42 +14,10 @@ from arq.connections import RedisSettings
 from .config import WorkerSettings as _WorkerSettings
 from .config import get_worker_settings
 from .dto import TimestepDTO
+from .file import acompress, atar, managed_file_system
 from .logging import configure_logging
 
 LOGGER = logging.getLogger(__name__)
-
-
-def compress(src: Path, target: Path):
-    LOGGER.info("Compressing %s", src)
-
-    img = cv2.imread(str(src))
-    cv2.imwrite(
-        str(target),
-        img,
-        params=(cv2.IMWRITE_TIFF_COMPRESSION, 5),
-    )
-
-    LOGGER.info("Compressed %s", src)
-
-
-async def acompress(src: Path, target: Path, executor: Executor):
-    loop = asyncio.get_event_loop()
-    await loop.run_in_executor(executor, compress, src, target)
-
-
-def tar(src: Path, target: Path):
-    LOGGER.info("Tarring %s", src)
-
-    target.parent.mkdir(parents=True, exist_ok=True)
-    with tarfile.open(target, "a") as tar:
-        tar.add(src, arcname=".")
-
-    LOGGER.info("Tarred %s", src)
-
-
-async def atar(src: Path, target: Path, executor: Executor):
-    loop = asyncio.get_event_loop()
-    await loop.run_in_executor(executor, tar, src, target)
 
 
 async def workflow(ctx: dict, input_data: dict):
@@ -79,29 +44,6 @@ async def workflow(ctx: dict, input_data: dict):
         await atar(temp_img_dir, temp_archive_path, pool)
 
         await s3._put_file(temp_archive_path, f"{settings.AWS_BUCKET_NAME}/{data.key}")
-
-
-@asynccontextmanager
-async def managed_file_system(
-    settings: _WorkerSettings,
-) -> AsyncGenerator[s3fs.S3FileSystem, None]:
-    client_kwargs = {}
-    if settings.AWS_REGION_NAME:
-        client_kwargs["region_name"] = settings.AWS_REGION_NAME
-
-    s3 = s3fs.S3FileSystem(
-        key=settings.AWS_ACCESS_KEY_ID,
-        secret=settings.AWS_SECRET_ACCESS_KEY,
-        endpoint_url=settings.AWS_ENDPOINT_URL,
-        client_kwargs=client_kwargs,
-        asynchronous=True,
-    )
-
-    session = await s3.set_session()
-
-    yield s3
-
-    await session.close()
 
 
 async def startup(ctx):
